fix(rbac): preserve explicit wildcard deny grant through dialog round-trip
The previous fix collapsed the synthetic kind-scoped wildcard
{resource_kind:"mcp", resource_id:"*"} to selectors:null in
grantsFromRole regardless of effect. sdkGrantsFromForm has no nil-deny
path — it only emits a deny entry when denySelectors.length > 0 — so a
deny rule with selectors:null was silently dropped on save. A user
attempting to clear that deny via the X chip also could not produce a
diff, leaving the stored deny untouched.

Restrict the wildcard collapse to allow grants only. Backend already
treats deny wildcards as explicit (scopeWildcardStaysExplicit test in
server/internal/authz/scopes_test.go), so the dialog now matches.

Also extract grantsFromRole/sdkGrantsFromForm/diffGrants into
roleGrantTransform.ts so they can be unit-tested without pulling in
React/react-query, and add round-trip coverage for both effects.
